refactor: Move has_index back to dataframes
